requirements.txt
asgiref==3.10.0
black==25.11.0
click==8.3.1
colorama==0.4.6
Django==5.2.8
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2==2.9.11
python-dotenv==1.2.1
pytokens==0.3.0
ruff==0.14.5
sqlparse==0.5.3
tzdata==2025.2
